COM-1917 move 401 interceptor in appcontainer
